+++ src/controllers/validator/did-array.ts
