+++ src/controllers/validator/did-document-id.ts
