+++ src/controllers/validator/jwt-proof.ts
