+++ src/test/validation/is_cheqd_w3c_credential.test.ts
@@ -1,4 +1,4 @@
-import { CheqdW3CVerifiableCredentialValidator } from '../../controllers/validator/cheqd_credential';
+import { CheqdW3CVerifiableCredentialValidator } from '../../controllers/validator/credential';
 import { CREDENTIAL_OBJECT } from '../constants';
 import { describe, it, expect } from '@jest/globals';
