+++ src/test/validation/is_create_did_service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from '@jest/globals';
-import { CreateDIDDocumentServiceValidator } from '../../controllers/validator/service_create_request';
+import { CreateDIDDocumentServiceValidator } from '../../controllers/validator/service-create-request';
 
 const createDIDDocumentServiceValidator = new CreateDIDDocumentServiceValidator();
